recorder/requirements.txt
opencv-python>=4.8.0
PyYAML>=6.0
